Core Data: Move EntitiesSavedState component out of editor
