+++ packages/core-data/src/components/entities-saved-states/entity-record-item.js
@@ -4,48 +4,40 @@
 import { CheckboxControl, PanelRow } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
-import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
 
 /**
  * Internal dependencies
  */
-import { store as editorStore } from '../../store';
-import { unlock } from '../../lock-unlock';
+import { STORE_NAME } from '../../name';
 import { getTemplateInfo } from '../../utils/get-template-info';
 
 export default function EntityRecordItem( { record, checked, onChange } ) {
 	const { name, kind, title, key } = record;
 
 	// Handle templates that might use default descriptive titles.
-	const { entityRecordTitle, hasPostMetaChanges } = useSelect(
+	const { entityRecordTitle } = useSelect(
 		( select ) => {
 			if ( 'postType' !== kind || 'wp_template' !== name ) {
 				return {
 					entityRecordTitle: title,
-					hasPostMetaChanges: unlock(
-						select( editorStore )
-					).hasPostMetaChanges( name, key ),
 				};
 			}
 
-			const template = select( coreStore ).getEditedEntityRecord(
+			const template = select( STORE_NAME ).getEditedEntityRecord(
 				kind,
 				name,
 				key
 			);
 
 			const { default_template_types: templateTypes = [] } =
-				select( coreStore ).getCurrentTheme() ?? {};
+				select( STORE_NAME ).getCurrentTheme() ?? {};
 
 			return {
 				entityRecordTitle: getTemplateInfo( {
 					template,
 					templateTypes,
 				} ).title,
-				hasPostMetaChanges: unlock(
-					select( editorStore )
-				).hasPostMetaChanges( name, key ),
 			};
 		},
 		[ name, kind, title, key ]
@@ -63,11 +55,6 @@ export default function EntityRecordItem( { record, checked, onChange } ) {
 					className="entities-saved-states__change-control"
 				/>
 			</PanelRow>
-			{ hasPostMetaChanges && (
-				<ul className="entities-saved-states__changes">
-					<li>{ __( 'Post Meta.' ) }</li>
-				</ul>
-			) }
 		</>
 	);
 }
