+++ packages/core-data/src/components/entities-saved-states/style.scss
