+++ packages/core-data/src/private-apis.ts
@@ -10,6 +10,13 @@ import {
 	useOnCollaboratorLeave,
 	useOnPostSave,
 } from './hooks/use-post-editor-awareness-state';
+import { getTemplateInfo } from './utils/get-template-info';
+import { getTemplatePartIcon } from './utils/get-template-part-icon';
+import {
+	EntitiesSavedStatesExtensible,
+	default as EntitiesSavedStates,
+} from './components/entities-saved-states';
+import { useIsDirty as useEntitiesSavedStatesIsDirty } from './components/entities-saved-states/hooks/use-is-dirty';
 import { lock } from './lock-unlock';
 import { retrySyncConnection } from './sync';
 import {
@@ -18,6 +25,11 @@ import {
 } from './utils/crdt-user-selections';
 
 const lockedApis = {
+	EntitiesSavedStates,
+	EntitiesSavedStatesExtensible,
+	getTemplateInfo,
+	getTemplatePartIcon,
+	useEntitiesSavedStatesIsDirty,
 	useEntityRecordsWithPermissions,
 	RECEIVE_INTERMEDIATE_RESULTS,
 	retrySyncConnection,
