+++ packages/core-data/src/style.scss
@@ -1,0 +1,1 @@
+@use "./components/entities-saved-states/style.scss" as *;
