+++ packages/core-data/src/utils/get-template-info.js
