+++ packages/core-data/src/utils/test/get-template-info.js
