+++ packages/core-data/tsconfig.json
@@ -8,15 +8,19 @@
 	},
 	"references": [
 		{ "path": "../api-fetch" },
+		{ "path": "../components" },
 		{ "path": "../block-editor" },
 		{ "path": "../blocks" },
 		{ "path": "../compose" },
 		{ "path": "../data" },
 		{ "path": "../deprecated" },
 		{ "path": "../element" },
+		{ "path": "../global-styles-engine" },
 		{ "path": "../html-entities" },
 		{ "path": "../i18n" },
+		{ "path": "../icons" },
 		{ "path": "../is-shallow-equal" },
+		{ "path": "../notices" },
 		{ "path": "../private-apis" },
 		{ "path": "../rich-text" },
 		{ "path": "../sync" },
