+++ packages/editor/src/components/post-card-panel/index.js
@@ -10,7 +10,10 @@ import {
 	privateApis as componentsPrivateApis,
 } from '@wordpress/components';
 import { close } from '@wordpress/icons';
-import { store as coreStore } from '@wordpress/core-data';
+import {
+	store as coreStore,
+	privateApis as coreDataPrivateApis,
+} from '@wordpress/core-data';
 import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
@@ -27,8 +30,8 @@ import {
 import { unlock } from '../../lock-unlock';
 import PostActions from '../post-actions';
 import usePageTypeBadge from '../../utils/pageTypeBadge';
-import { getTemplateInfo } from '../../utils/get-template-info';
 const { Badge: WCBadge } = unlock( componentsPrivateApis );
+const { getTemplateInfo } = unlock( coreDataPrivateApis );
 
 /**
  * Renders a title of the post type and the available quick actions available within a 3-dot dropdown.
