+++ packages/editor/src/store/private-actions.js
@@ -2,9 +2,9 @@
  * WordPress dependencies
  */
 import { store as coreStore } from '@wordpress/core-data';
+import { store as blockEditorStore } from '@wordpress/block-editor';
 import { __, _x, sprintf } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
-import { store as blockEditorStore } from '@wordpress/block-editor';
 import { store as preferencesStore } from '@wordpress/preferences';
 import { addQueryArgs } from '@wordpress/url';
 import apiFetch from '@wordpress/api-fetch';
@@ -126,130 +126,6 @@ export const hideBlockTypes =
 		registry
 			.dispatch( preferencesStore )
 			.set( 'core', 'hiddenBlockTypes', [ ...mergedBlockNames ] );
-	};
-
-/**
- * Save entity records marked as dirty.
- *
- * @param {Object}   options                        Options for the action.
- * @param {Function} [options.onSave]               Callback when saving happens.
- * @param {object[]} [options.dirtyEntityRecords]   Array of dirty entities.
- * @param {object[]} [options.entitiesToSkip]       Array of entities to skip saving.
- * @param {Function} [options.close]                Callback when the actions is called. It should be consolidated with `onSave`.
- * @param {string}   [options.successNoticeContent] Optional custom success notice content. Defaults to 'Site updated.'.
- */
-export const saveDirtyEntities =
-	( {
-		onSave,
-		dirtyEntityRecords = [],
-		entitiesToSkip = [],
-		close,
-		successNoticeContent,
-	} = {} ) =>
-	( { registry } ) => {
-		const PUBLISH_ON_SAVE_ENTITIES = [
-			{ kind: 'postType', name: 'wp_navigation' },
-		];
-		const saveNoticeId = 'site-editor-save-success';
-		const homeUrl = registry
-			.select( coreStore )
-			.getEntityRecord( 'root', '__unstableBase' )?.home;
-		registry.dispatch( noticesStore ).removeNotice( saveNoticeId );
-		const entitiesToSave = dirtyEntityRecords.filter(
-			( { kind, name, key, property } ) => {
-				return ! entitiesToSkip.some(
-					( elt ) =>
-						elt.kind === kind &&
-						elt.name === name &&
-						elt.key === key &&
-						elt.property === property
-				);
-			}
-		);
-		close?.( entitiesToSave );
-		const siteItemsToSave = [];
-		const pendingSavedRecords = [];
-		entitiesToSave.forEach( ( { kind, name, key, property } ) => {
-			if ( 'root' === kind && 'site' === name ) {
-				siteItemsToSave.push( property );
-			} else {
-				if (
-					PUBLISH_ON_SAVE_ENTITIES.some(
-						( typeToPublish ) =>
-							typeToPublish.kind === kind &&
-							typeToPublish.name === name
-					)
-				) {
-					registry
-						.dispatch( coreStore )
-						.editEntityRecord( kind, name, key, {
-							status: 'publish',
-						} );
-				}
-
-				pendingSavedRecords.push(
-					registry
-						.dispatch( coreStore )
-						.saveEditedEntityRecord( kind, name, key )
-				);
-			}
-		} );
-		if ( siteItemsToSave.length ) {
-			pendingSavedRecords.push(
-				registry
-					.dispatch( coreStore )
-					.__experimentalSaveSpecifiedEntityEdits(
-						'root',
-						'site',
-						undefined,
-						siteItemsToSave
-					)
-			);
-		}
-		registry
-			.dispatch( blockEditorStore )
-			.__unstableMarkLastChangeAsPersistent();
-
-		Promise.all( pendingSavedRecords )
-			.then( async ( values ) => {
-				if ( onSave ) {
-					await onSave();
-				}
-				return values;
-			} )
-			.then( ( values ) => {
-				if (
-					values.some( ( value ) => typeof value === 'undefined' )
-				) {
-					registry
-						.dispatch( noticesStore )
-						.createErrorNotice( __( 'Saving failed.' ) );
-				} else {
-					registry
-						.dispatch( noticesStore )
-						.createSuccessNotice(
-							successNoticeContent || __( 'Site updated.' ),
-							{
-								type: 'snackbar',
-								id: saveNoticeId,
-								actions: [
-									{
-										label: __( 'View site' ),
-										url: homeUrl,
-										openInNewTab: true,
-									},
-								],
-							}
-						);
-				}
-			} )
-			.catch( ( error ) =>
-				registry
-					.dispatch( noticesStore )
-					.createErrorNotice(
-						`${ __( 'Saving failed.' ) } ${ error }`
-					)
-			);
 	};
 
 /**
