+++ packages/editor/src/store/selectors.js
@@ -12,7 +12,10 @@ import { addQueryArgs, cleanForSlug } from '@wordpress/url';
 import { createSelector, createRegistrySelector } from '@wordpress/data';
 import deprecated from '@wordpress/deprecated';
 import { store as blockEditorStore } from '@wordpress/block-editor';
-import { store as coreStore } from '@wordpress/core-data';
+import {
+	store as coreStore,
+	privateApis as coreDataPrivateApis,
+} from '@wordpress/core-data';
 import { store as preferencesStore } from '@wordpress/preferences';
 
 /**
@@ -26,10 +29,10 @@ import {
 	AUTOSAVE_PROPERTIES,
 } from './constants';
 import { getPostRawValue } from './reducer';
-import { getTemplatePartIcon } from '../utils/get-template-part-icon';
 import { unlock } from '../lock-unlock';
-import { getTemplateInfo } from '../utils/get-template-info';
 import { getDeviceTypeByCanvasWidth } from '../utils/device-type';
+
+const { getTemplateInfo, getTemplatePartIcon } = unlock( coreDataPrivateApis );
 
 /**
  * Shared reference to an empty object for cases where it is important to avoid
