+++ packages/editor/src/style.scss
@@ -17,7 +17,6 @@
 @use "./components/document-outline/style.scss" as *;
 @use "./components/document-tools/style.scss" as *;
 @use "./components/editor-interface/style.scss" as *;
-@use "./components/entities-saved-states/style.scss" as *;
 @use "./components/error-boundary/style.scss" as *;
 @use "./components/global-styles/style.scss" as *;
 @use "./components/global-styles-sidebar/style.scss" as *;
