+++ packages/editor/src/utils/index.js
@@ -5,4 +5,3 @@ import mediaUpload from './media-upload';
 
 export { mediaUpload };
 export { cleanForSlug } from './url.js';
-export { getTemplatePartIcon } from './get-template-part-icon';
